Continuing PE Dev Testing
Added a thermal modeling aspect to testPEConvSubmodule.slx using the pulsed power device loss approximation to model the thermal response for a single device.

Created a plecs folder that holds a model of 2L VSI with thermal simulations to verify the pulsed-power device loss approximation.

Also added the temperature-dependencies for rdson and diode resistance in peDevClass.